challenge bonus

+++ movieHandlers.js
@@ -123,20 +123,24 @@ const updateMovie = (req, res) => {
 
 const deleteMovie = (req, res) => {
   const id = parseInt(req.params.id);
-
-  database
-    .query("delete from movies where id = ?", [id])
-    .then(([result]) => {
-      if (result.affectedRows === 0) {
-        res.status(404).send("Not Found");
-      } else {
-        res.sendStatus(204);
-      }
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error deleting movie");
-    });
+  console.log("req.payload", req.payload);
+  if (id === req.payload) {
+    database
+      .query("delete from movies where id = ?", [id])
+      .then(([result]) => {
+        if (result.affectedRows === 0) {
+          res.status(404).send("Not Found");
+        } else {
+          res.sendStatus(204);
+        }
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send("Error deleting movie");
+      });
+  } else {
+    res.status(403).send("Not Found");
+  }
 };
 
 module.exports = {
